backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements/requirements.in
#
passlib==1.7.1
pyjwt==1.7.1
requests==2.2.1
